requirements.txt
pymysql
Flask
Flask-Ext
django
django-crispy-forms
scrapy 
mysql-connector-python
absl-py==0.9.0
astor==0.8.1
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
cssselect==1.1.0
feedparser==5.2.1
gast==0.3.3
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
joblib==0.16.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lxml==4.5.2
Markdown==3.2.2
numpy==1.19.1
Pillow==7.2.0
pkg-resources==0.0.0
protobuf==3.12.4
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
requests-file==1.5.1
scikit-learn==0.23.1
scipy==1.5.2
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
threadpoolctl==2.1.0
tldextract==2.2.2
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
spacy
tika
nltk